chore(docs): Optimized array indexing in the log monitor widget's rendering logic
# Project State

## Current Focus
Optimized array indexing in the log monitor widget's rendering logic

## Context
The change improves performance by removing unnecessary type casting during array indexing operations in the log monitor widget's rendering logic.

## Completed
- [x] Removed redundant `as usize` cast in array index calculation
- [x] Simplified index calculation for better performance

## In Progress
- [ ] None

## Blockers
- None

## Next Steps
1. Verify performance impact with benchmarks
2. Review for any potential edge cases in the rendering logic
---
category: chore
scope: docs
---

+++ examples/_cookbook/log_monitor.rs
@@ -262,7 +262,7 @@ impl Widget for LogMonitor {
             }
         }
         for x in 1..w - 1 {
-            let idx = ((h - 1) * w + x) as usize;
+            let idx = (h - 1) * w + x;
             p.cells[idx].bg = t.surface_elevated;
         }
